Re-integrate tailwind
