+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "qwiket-findexar",
+  "name": "q2",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -9,6 +9,9 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "next": "15.1.6",
     "@clerk/nextjs": "^6.10.6",
     "@clerk/themes": "^2.0.1",
     "@stripe/react-stripe-js": "^3.1.1",
@@ -22,11 +25,8 @@
     "iron-session": "^8.0.1",
     "isbot-fast": "^1.2.0",
     "material-icons": "^1.13.12",
-    "next": "15.1.6",
     "next-themes": "^0.3.0",
-    "react": "19.0.0",
     "react-chartjs-2": "^5.2.0",
-    "react-dom": "19.0.0",
     "react-icons": "^5.3.0",
     "react-intersection-observer": "^9.15.1",
     "react-markdown": "^9.0.3",
@@ -41,19 +41,16 @@
     "use-stripe-subscription": "^1.0.6"
   },
   "devDependencies": {
-    "@tailwindcss/typography": "^0.5.16",
+    "typescript": "^5",
     "@types/node": "^20",
-    "@types/react": "19.0.8",
-    "@types/react-dom": "19.0.3",
-    "autoprefixer": "^10.4.20",
-    "eslint": "^8",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
+    "postcss": "^8",
+    "tailwindcss": "^3.4.1",
+    "eslint": "^9",
     "eslint-config-next": "15.1.6",
-    "postcss": "^8.5.1",
-    "tailwindcss": "^4.0.3",
-    "typescript": "^5"
-  },
-  "overrides": {
-    "@types/react": "19.0.8",
-    "@types/react-dom": "19.0.3"
+    "@eslint/eslintrc": "^3",
+    "@tailwindcss/typography": "^0.5.16",
+    "autoprefixer": "^10.4.20"
   }
 }
