--- tailwind.config2.ts
@@ -1,34 +1,0 @@
-import type { Config } from "tailwindcss";
-
-
-const config: Config = {
-    darkMode: 'class',
-
-    content: [
-        "./lib/**/*.{js,ts,jsx,tsx,mdx}",
-        "./components/**/*.{js,ts,jsx,tsx,mdx}",
-        "./app/**/*.{js,ts,jsx,tsx,mdx}",
-    ],
-    theme: {
-        typography: (theme: any) => ({}),
-        extend: {
-            typography: () => ({
-                dark: {
-                    css: {
-                        color: 'white',
-                    },
-                },
-            }),
-            backgroundImage: {
-                "gradient-radial": "radial-gradient(var(--tw-gradient-stops))",
-                "gradient-conic":
-                    "conic-gradient(from 180deg at 50% 50%, var(--tw-gradient-stops))",
-            },
-        },
-    },
-
-    plugins: [
-        require('@tailwindcss/typography'),
-    ],
-};
-export default config;
